make swr global
