+++ javascript/backend-api-routes_products/pages/_app.js
@@ -1,10 +1,29 @@
+import useSWR, { SWRConfig } from "swr";
+
 import GlobalStyle from "../styles";
+
+const fetcher = async (url) => {
+  const res = await fetch(url);
+  if (!res.ok) {
+    const error = new Error("An error occurred while fetching the data.");
+    error.info = await res.json();
+    error.status = res.status;
+    throw error;
+  }
+  return res.json();
+};
 
 export default function App({ Component, pageProps }) {
   return (
     <>
-      <GlobalStyle />
-      <Component {...pageProps} />
+      <SWRConfig
+        value={{
+          fetcher,
+        }}
+      >
+        <GlobalStyle />
+        <Component {...pageProps} />
+      </SWRConfig>
     </>
   );
 }
