+++ javascript/backend-api-routes_products/pages/products/index.js
@@ -1,25 +1,24 @@
 import useSWR from "swr";
 import { SWRConfig } from "swr";
 
-const url = "/api/products";
-const fetcher = async (url) => {
-  const res = await fetch(url);
+// const fetcher = async (url) => {
+//   const res = await fetch(url);
 
-  // If the status code is not in the range 200-299,
-  // we still try to parse and throw it.
-  if (!res.ok) {
-    const error = new Error("An error occurred while fetching the data.");
-    // Attach extra info to the error object.
-    error.info = await res.json();
-    error.status = res.status;
-    throw error;
-  }
+//   // If the status code is not in the range 200-299,
+//   // we still try to parse and throw it.
+//   if (!res.ok) {
+//     const error = new Error("An error occurred while fetching the data.");
+//     // Attach extra info to the error object.
+//     error.info = await res.json();
+//     error.status = res.status;
+//     throw error;
+//   }
 
-  return res.json();
-};
+//   return res.json();
+// };
 
 export default function ProductsList() {
-  const { data, error, isLoading } = useSWR(`api/products/`, fetcher);
+  const { data, error, isLoading } = useSWR(`/api/products/`);
 
   if (error) return <div>{error.message}</div>;
   if (isLoading) return <div>loading...</div>;
